Fix corrupt model name
The lens correction plugin is showing a corrupted model for me:

EOS DIGITAL REBEL XTBEL XTi

which looks like we had "CANON EOS DIGITAL REBEL XTi" and copied over it with
the first 20 characters of "EOS DIGITAL REBEL XTi" without the NULL terminator.

strncpy will not append a NULL terminator if the string to be copied
exceeds the length specified. Fix this by always adding a NULL terminator
to the last character when copying the maker and model.

We only copy in the first 20 characters of the maker and model, even
though the destination strings are 32 bytes. Avoid hardcoding the length
by using sizeof().