Recursive import and ignoring of JPEG
The file chooser got two new check boxes:
- recursive import
- ignoring of jpeg

The first controls if subdirectories are processed, too. If selected
they go into their own film rolls.

The latter controls if JPEG files are to be imported. This is a hack to
suit raw+jpeg photographers who have their files in one directory. Since
there is no point in importing JPEG files if one has RAW this option
helps.

I also added .xmp to the test for .dt/.dttags as we use .xmp now.

FIXME:
The recursive import has one annoying bug: The lists of recent film
rolls and all film rolls are not updated with the film rolls of
subdirectories. Only the top level which was selected in the open dialog
is shown. As soon as any film roll is selected the lists are finally updated.