New variable in the export path.
I've added a new variable to create the path for the "file on disk"
export.

PATCH ATTACHED !!!

$(ROLL_NAME) - roll of the input image

This variable was added in order to better organize the disk folder
tree.

Because each picture is inside its own folder that has the name of the
roll, the picture developed can be exported to another folder (in
another layer) with the same name as the folder of the source image.

Eg.:

.../Pictures/sources/2010.01.01/IMG_123.NEF

can be exported in

$(PICTURES_FOLDER)/developed/$(ROLL_NAME)/$(FILE_NAME).jpg

equal to .../Pictures/developed/2010.01.01/IMG_123.jpg

I think it's very usefull !

I hope that you will want to use it.

Let me know !

Bye Bye !!!