DT patch - don't squash ca_correct preset
hanatos -

First, thanks for a great piece of software!  It produces better images
than I ever managed with UFraw.

Second, I've been using AMaZE for demosaicing, with the CA autocorrect
because my lenses aren't the greatest.  However, when I save it as a
preset, the flag is saved in the preset blob in the SQLite database but
it's not restored when you activate the preset.

A little digging around in rawimport.c turned up the reason.  When the
preset loads, it calls gui_update(), which updates the demosaicing combo
box.  The combo box callback shows the ca correct checkbox - and resets
the flag in p->fill1.  If we make those two steps conditional on the
current demosaic mode, the problem goes away.

Anyways, please find attached a patch against git HEAD.  I look forward
to continuing to use, and occasionally contribute to, darktable!

Regards,
Brian