Add title, creator and rights to the metadata on the left
Also make sure that the result of dt_metadata_get() is always freed
when it is no longer needed. I assume that there is only a single entry
in the metadata table per image. When we can assume that everyone has
at least GTK 2.28 we should use g_list_free_full().