Add orientation to database
This requires sqlite > 3.1.3 since earlier version can't open tables
which were altered. If the database was not converted but created
directly with the orientation column it should even work with older
versions.
As now all the data in dt_image_t is stored in the database we can keep
the exif_inited flag when caching images, so that it's not necesarry to
reparse the exif data upon loading. Entering darkroom feels even a bit
faster now. :-)