Add highlighting of over/under exposed pixles
I just put it in the end of the pipe for it had to be anywhere. In the
moment overexposed pixels are turned red, underexposed ones blue. I
guess a color picker/slider to change these colors would be nice. I'm
also not sure if the test to determine if a pixel is to be colored is so
great: I just do "[r|g|b] [<=|>=] threshold", so it's the same for all
three channels.