Film view UI bug patch
I've got lots of film rolls in the database, so I have to scroll down to
get to some of them.  When I select one, the treeview jumps back to the
top instead of staying where I left it.

If this is intended behavior, then let me just say it drives me nuts.
If it's a bug, then please find attached a patch to fix it.

On a related note - is the mailing list an appropriate place to bring up
these issues?  Or should I use trac - I didn't know it was available
until yesterday.  (=

Brian

--
Brian Teague
teague@wisc.edu
The Laboratory for Molecular and Computational Genomics

Behold the turtle: he makes progress only when he sticks his neck out.
               -- James B. Conant