ticket #23: Fix this once and for all.
When importing folders/files, make sure that the list of all filmrolls
is updated, too, not only the recently used ones. While the latter is
done in dt_film_open(), the first has to be handled in the button
callbacks. This has also be done when opening an image from command
line.