ticket #105: Debug helpers for sqlite
debug.h defines wrappers around some sqlite3-methods which make sure
that the return value is checked. If compiled in Debug mode an assert()
will be inserted so that dt crashes whenever something goes wrong. If
_DEBUG is not defined it will just print an error message to stderr.

In order to make use of this simply replace the following sqlite3 calls:
- sqlite3_exec -> DT_DEBUG_SQLITE3_EXEC
- sqlite3_prepare_v2 -> DT_DEBUG_SQLITE3_PREPARE_V2
- sqlite3_bind_int -> DT_DEBUG_SQLITE3_BIND_INT
- sqlite3_bind_double -> DT_DEBUG_SQLITE3_BIND_DOUBLE
- sqlite3_bind_text -> DT_DEBUG_SQLITE3_BIND_TEXT
- sqlite3_bind_blob -> DT_DEBUG_SQLITE3_BIND_BLOB

Currently dt will crash when compiled in debug mode due to some SQL
strangeness.