Fix deletion of images with duplicates
1. If an image without duplicates is deleted, well, it's just gone.
2. If however there are duplicates, then only the sidecar files are
   removed and the actual image on disk is not touched. Only when the
   delete operation is executed for the last of the duplicates GOTO 1

FIXME: If an image with duplicates is removed, and these duplicates were
created after the one which got removed (i.e. the imgid of the
duplicates is bigger), then the xmp filenames of these images will
change. In the best case they will be decreased by 1 so that the xmp
file of the image with the biggest id will be on disk twice. In the
worst case, some xmp files will be decreased, overwriting the sidecars
of others. Not good. We either have to keep track of the version number
of every image in the db so that it will never change. Or we could
delete+recreate all the xmp files of affected images. Boy, that's a long
one, isn't it? And you read all of this? Let's hope you also understood
what I was writing. ;-) kthxby