Centralize and expand performance reporting
The output from darktable -d perf would show only the wall (clock) time
used by the plugins.  The CPU time used is also useful for performance
analysis, sometimes more so, as it is disturbed less if something else
is running.  This change centralizes all the time reporting in one place
which reports both the clock and CPU times.  It also disables all time
calculations if -d perf is not turned on.