5 star rating
Hej, guys,

I tried to understand the code and wrote a patch implementing a extended
star rating system. With this patch it is possible to:

- rate a picture with starts 0 .. 5 (with mouse or key accelerators 0 ..
5, 0 stars can still be achieved by pressing "1" again)
- rate a picture as "rejected" (NOT with mouse, but key acc. "Delete")
rejected pictures are marked with one red star (only border) Is this
intuitive? it was the best way that came to my mind...

The rating is written to xmp files correctly (range -1 .. 5) according
to the Adobe xmp specification found here:
http://www.adobe.com/content/dam/Adobe/en/devnet/xmp/pdfs/XMPSpecificationPart1.pdf

In the database, the rating is saved as 0 .. 6 where 0 .. 5 represent
the number of stars and 6 being the rejected status. Rating works in
lighttable mode, in filmstrip and in the "zoom" mode (activate by using
"z").

TODO:
- think about a nice way to mark a picture as "rejected" by mouse (and
assign stars again)
- extend and adjust the "show" option in the top panel, maybo do not
show rejected pictures by default (so differ between "all" and "all +
rejected")

Would please somebody have a look at this? It has been some time since I
coded the last time...
Cheers,
Simon