sharpen: Optimize 2D gaussian blur as two 1D blur operations
Due to the nature of the gaussian blur operation, this reworking
produces identical results (modulo rounding) as the full operation, but
can run significantly faster.  For the default radius of 2, the
difference is not large (5+5 steps plus temporary store/load overhead
instead of 5*5), but it grows much larger with increased radii.