overexposed: Mark pixels if any instead of if all
The overexposed plugin marks pixels only if all of the channels are
above the upper limit.  Overexposure flaws can cause significant color
shifting if only one or two channels are blown out.  This patches
switches the test logic to mark pixels if any of the channels are over
the limit.