Permutohedral: Several small optimizations
The biggest improvement is using a bit mask instead of modulus for hash
table calculations.  This small change improves the performance of the
whole algorithm by about 15%.

Eliminating the extraneous myrank/mygreedy variables in splat simplifies
the code slightly, and results in a small but repeatable performance
improvement in captive benchmarks.

Using int instead of char or short for intermediate calculations in
splat similarly results in a small improvement.  The smaller arrays save
trivial amounts of memory, but cause extra operations to truncate the
integer intermediate values down to fit.

Sum total of these improvements is about a 20% performance boost for the
tonemap and bilateral filters.