colorcontrast: change steepness of a* and b* curves
Changing the steepness of a* and b* curves in Lab color space is a
popular method to boost color contrast in the image. This plugin allows
to control the effect with only two sliders. They have user-friendly
labels "green vs magenta" and "blue vs yellow".

These two values are the tangents of straight symmetric curves for the
transforms in a* anb b* channels respectively.