A very big refactoring blob :) central changes is:
The view should take care of what to show or not and does
so by using plugins libs/*.c which have flags to decide where
to show up, this can later be refactored to be user configurable.

New flags are added and all plugins are updated.

The view enter() and leave() should not control what parts of UI to
show as todays hardcoded widgets in gui/gtk.c im working on moving
this to plugins and therefor ui is kind of broken by this commit
but its intended.