Fix file path when importing images from the camera
This patch introduces two new functions to common/utility.h - One is to retreive
the users home directory and the other one to fix a given file path by replacing
a prepended tilde with the correct home directory.

Those functions are called in the camera_import_dialog as well as in the
camera_jobs to allow path settings like "~/img" to point to the right directory
instead of being created in the users home directory.