fix deep blue issue with negative L
Deep blue colors produce negative L values.
This patch fixes behavior of tonecurve for them
by linear approximation of curve for L values in range
of (-inf,1.0f)

Since it is necessary to preserve continuity
and because of curve calculation errors for small values,
linear approximation is done for pixels where L < 1.0f.