opencl: revamp ROUNDUP mechanism
Width and height of global work group are no longer rounded up to
a multiple of 4 (fixed) but to a user adjustable multiple of
new config parameter "opencl_size_roundup". Default value is 16.
The original value of 4 was found to be not optimal. The new default
can lead to significant speed-up in opencl processing (esp. export).
Users can easily test if other values give even better results for them.