Make sure tags are written to sidecar files
There were a few places in the code where the xmp files were not synched
when a tag was added/removed. To fix this once and for all I put the
synching into the functions that utility functions that add/remove the
tag. People using these helpers should not care about all the details
that have to be done anyways.