The UI is updated when the collection is changed
When the collection is changed, the signal DT_SIGNAL_COLLECTION_CHANGED is
raised. We connect to it to redraw the UI then.
TODO: this is not optimal, as it also redraws the UI when we select a
filmroll or the filter is changed. It should be updated when there is
a change in the filmrolls table or in the number of images.