interpolation: add a SSE specialized 4 component pixel interpolator
This assumes pixels to be interpolated the same for all components.
So we can happily save 2/3 of current kernel computation. Plus of
course we compute the 4 components in single pass.

Ported clipping.c to this new API. lens.c cannot benefit from this
as it theoritically interpolates each color plane independently

For my test image:
lanczos3: 0.164s -> 0.059s