dt_cache_init(): don't activate cache_mask if no cacheline optimization.
Since cache->optimize_cacheline is currently forced to FALSE because the code
is not prepared to deal with LRU lists, there is no need to set the cache_mask
field either.

Note that this change hides the problem I was having on the 10x branch
with darktable complaining:

	[cache] failed to find a free spot for new data!

whereas the cache was far from being full:

	[cache] full entries:
	[cache] bucket 0 holds key 12 with locks r 0 w 0
	[cache] bucket 1 is empty with locks r 0 w 0
	[cache] bucket 2 is empty with locks r 0 w 0
	[cache] bucket 3 holds key 18 with locks r 0 w 0
	[cache] lru entries:
	[cache] bucket 0 holds key 12 with locks r 0 w 0
	[cache] bucket 3 holds key 18 with locks r 0 w 0

but it does not explain by itself why the dt_cache_read_get() routine failed
to find the free spot.

Still, there's really no need for this cache_mask setting when the cacheline
optimization is turned off.