interpolation: add meta index for accessing resampling plan easily
For future OpenMPization of the main processing loop, we must be able
to access all arrays (kernel, index, length) based on line number as
each processing thread will start wherever OpenMP decides to split
the loop.

- add metaindex array of triplets (length, kernel, index)
- use that for the vertical plan
- use the meta index for all vertical indexes in the main processing
  loop

Hopefully, this seems to be the only required change for a future
OpenMPization of the loop. Stay tuned !