Fix handling of xmp files.
Updating the film_id directly after moving one image of a set of duplicates
caused dt_image_path_append_version() to return wrong results, as the
database was already partly changed for the set.

The fix is first moving all xmps of the set and then updating the film_ids.