Allow to quit dt when in map mode
WARNING: this is an ugly hack. Currently map mode gets rid of the center
widget and hands the space (and all the control) over to the map widget.
This has the drawback that the global expose function no longer runs and
as a consequence the program is never terminated. I circumvented that by
just changing to lighttable mode in the quit callback in case we are in
map mode. As I said: not nice, but it works ...