Add popup for offset calculation
Usage: select an image which shows a time, click on the calculate offset
button in the geolocation module and enter the time shown on the image
into the popup. confirm with enter or click the ok button. the result
should be an updated offset in the entry box in the geolocation module.
next you can select some images and apply that offset.