Add a (partial) copy of colord-gtk to our tree
This are just the files we need to get a color profile from colord. The
reason to put it into our tree is that the version shipped with
distributions is linked against gtk3 which breaks stuff for us. Since
the parts that we need also run with gtk2 this seems to be the best
solution.
WARNING: the version of this commit doesn't compile, fix is in the next
commit to keep stuff separated!