Add option to change the limit of shown images on map
Also try to be smart which images to select: prefer those that are
nearest to the center of the map. This will introduce some changes to
the stacking order when panning. I might throw some more SQL magic at it
later (wrap the current query into an "select ... order by id")