Nicer drawing of group borders
Since thumbnails could overlap group borders they looked ugly in the
corners. Now we first draw all the thumbnails and iterate over the list
of images a 2nd time to draw the borders. This somehow delays the
borders a little but they look way better, so I guess it's worth it. I
also changed the red border to a more orange color. Thanks Pascal for
the push. :D