Allow smaller window heights
With the old min height of the 500 of the center drawing area the min
height of the window was 622 on my system which makes dt unusable on
some small screens. I reduced the min height of the center area to 200
which is obviously too small to do anything useful but this makes it at
least possible for the user to scale the window down.