dt_get_num_atom_cores: Add *BSD support
We first use the hw.model sysctl and check if it contains "Atom". Then
we query the hw.ncpu sysctl to get the number of cores.

This was tested on FreeBSD. The code is believed to work on DragonFly,
NetBSD and OpenBSD, according to sysctl(3) for DragonFly and OpenBSD and
sysctl(7) for NetBSD.