Correctly reimport pictures that got removed recently
When removing images from DB we set a bit in flags and then write the xmp
files. When reimporting the film roll right after removing it that flag
sometimes survives (either directly in the DB or over the xmp file, I
haven't checked). This should no longer happen. I hope.