Add support for adding a style while exporting.
This makes it possible to apply a specific style for some export
location like a watermark when exporting to Picasa for example.

At the same point it is more general as any style can be applied
at export time. This could be useful too when exporting for
printing where a different kind of sharpness is needed.