Add new dialog to copy only parts of the history.
This is activated by default on the lightable were ctrl-c and copy button
does a partial copy now. The dialog has a "copy all" button which should
be easy to use to get the previous behavior. In the film strip the ctrl-c
is mapped to the partial history copy (with the dialog) and a new action
mapped on ctrl-shift-c can be used to copy all history.