Fix film strip offset computing.
This may be due to a bug in SQlite as the request of the form:

   select rowid from (select * from images);

returns a suite of empty lines instead of the rowid. This was used
to compute the offset of the filmstrip and so the selected image
was never properly put in the center.

The workaround is to do part of the computation directly in C.