thumbnails: make last commit optional, default to high-quality.
(except for weak systems, where i switch it on by default)

it's not exposed in the gui, as the option is read in two different places
which have to match (or else we die). so it could be cached in darktable.something
instead, to ensure it's only read at startup.