Move colorzones definitions to header file.
This is needed for the lightroom import. At the same time we are
adding two 2 zones from 6 to 8. This gives a better precision and
is at the same time more compatible with what lightroom does and so
will ease the implementation of the import.
(cherry picked from commit 217a04c7e320f2bd48882fce91c1387edfe9b8fa)