Once again, fix build issue with libintl, GCC and *BSD
I still don't understand what's going on here, but this workaround seems
enough for now.

In the end, here's what we do on *BSD systems:
    o  clang: don't link with libintl
    o  GCC 4.2: don't link with libintl
    o  GCC 4.3+: explicitely link with libintl

This was tested on FreeBSD 9 & 10, with clang 3.2 (on FreeBSD 10 only,
using the default compiler), GCC 4.2 (base compiler) and GCC 4.6 (from
the ports tree).

The workaround is used on other BSD too, but honestly, I haven't check
if it was needed or if it works.