opencl: work around a bug in some OpenCL drivers
We add a call to clFinish() just before blending kernels are started.
Without this synchronization point, reads to dev_in would often not return
the correct value for some OpenCL drivers (namely AMD).
This bug also depends on the module after which blending is called. One of the
affected ones is sharpen.