Add an absolute translation for the watermark.
In this mode (new absolute toggle in the GUI), the watermark position is
not computed against the size/crop/scaling of the image but it is absolute.
This ensures that the watermark is always at the same position, this mode
is a usability fix when used into a preset or part of an export style. Indeed
in this case we do not control the position of the watermark in every picture
and it is expected to be at the same position.

Note that when the absolute mode is not used the behavior is the same
as before. This makes this change upward compatible.