Rework the watermark rendering.
This version does not use constant literal anymore. It introduces three
mode for the watermark placement.

image :          this is as before, alignement/scale are based on the
                 image width and depends on the image orientation.

larger border  : the alignement/scale are computed based on the larger
                 border.

smaller border : the alignement/scale are computed based on the smaller
                 border.

The last two modes can be used in styles are independant of the actual
image orientation. This new implementation try to be cleaner by using
most of the time original image coordinate. It should be easier to
maintain.