Fix Lightroom flip orientation.
Do not try to use the flip from the crop iop as it seems to be not
working in some cases. Instead of beeing smart we now simply use the
orientation module to rotate the image if needed. This anyway simplify
the code. Also make sure that the atof() is locale independant as
lightroom always store the float values with a dot as decimal
separator.