blending: add option to toggle polarity of drawn mask
this is needed to effectively extend a parametric mask by including or
excluding specific parts of an image.

to include parts: switch combobox combine to "inclusive" and make sure
polarities of drawn mask and parametric mask channels are set to negative.

to exclude parts: switch combobox combine to "exclusive" and make sure
polarities of drawn mask and parametric mask channels are set to positive.

use "inclusive & inverted" or "exclusive & inverted" to get an inverted mask
if needed.

the "toggle polarity" button (circle with black&white) can be used to quickly
toggle between inclusive and exclusive mode.