In full view on the lighttable, wrap around.
That is when pressing right arrow on the last image in a row we move
to the next image on the following row. Same thing for left arrow on
the first image of a row.

This makes reviewing all images more friendly.
(cherry picked from commit 823ec165a802d7da23820e5d96002e811bd1777a)