Removes support for deleting images from camera.
This option has proved to be dangerous and should anyway never be used.
It is far safer to keep the pictures on the sd/xd/cf card and letting
time to develop and put the pictures on a safe storage (RAID, external
storage service...). Then when a card is to be reused it can be formatted
on the camera.
(cherry picked from commit b462992410cabc0502fc72e9d24698da5864d5ae)