fix use of color picker and auto-exposure in exposure module
this is needed as exposure now allows multi-instance combined with masks.
the color picker from parametric masks needs to be differentiated from
the color picker in the auto-exposure functionality. in all cases parametric
masks sets flag request_color_pick to -1 whereas a module's color picker always
sets this flag to +1.