When apply a style to a duplicate make sure the collected images are updated.
Now that the collected images into the lighttable are stored separately
into a memory table, we want to ensure that this collection is properly
updated when applying a style to a duplicate. we were not doing this
and the duplicated image was not visible until some filters were changed.
(cherry picked from commit 1a1bdc4217f4edc1fc93f69a47918f08f5f8fe00)