Fix missing format tags for images with XMP
The problem is that we now go through some strange duplicate detection
code which re-reads the xmp file, which in turn removes the format tags
again since these are not stored in the xmp. For now I removed removal
of old metadata when reading an xmp file. I am neither sure if it was the
wisest choice to do so in the first place nor if this is any better. But
at least the tags work again.