Fix #9567: Only export the current image from dr
While this fixes the bug at hand it also removes a feature: selecting
arbitrary images in filmstrip and exporting them. But since darkroom is
meant for single image handling the bug is more important than such a
feature. The alternative would be to always reset the selection when
entering darkroom (currently this is only done when < 2 images are
selected) but that would hurt more people as selections would never
survive a trip to darkroom.

If this turns out to hurt more people than it helps: revert