Add some checks to ensure that a local copy can be created.
Without this checks dt could be in an inconsistent state with
no local copies created but the local copy status set on the
images. Furthermore, in this case the local copy status cannot
be removed.

At this stage we cannot create new messages so we just exit
silently. It is still better as the local copy status won't
be displayed on the lighttable.