tonecurve revamped
we no longer clamp the a and b channel when using the corresponding curves.
instead an extrapolation is applied at both sides of the curve. this fixes
possible hue banding and artifacts mainly for images with certain highly saturated
colors (think of blue illuminated buildings at night, stage photos etc.)

as usual existing history stacks are kept untouched - if you want to take
advantage of the improvements for those you need to reset the tonecurve module.