shadows&highlights: a more reasonable low approximation limit for color adjustment
this prevents within s&h some of the typical artifacts in highly saturated blues
(black pixels). the original threshold for low approximation was insanely high.
legacy history stacks are processed like with the old value in order to keep results
consistent. you need to reset the module in these cases to take advantage of the improvement.