Have a msgctxt for 'export' where it's a verb.
The reason for this is that in english, 'export' is a verb ('to export')
as well as a noun ('an export').  For example in Swedish, the verb is
'exportera' while the noun is 'export', and therefore, a separation of
some kind is needed.

Personally, I would have preferred having the contexts 'verb' and 'noun'
for this case, but this is not how things have been done so far in dt.