Removed check for ambiguous return values for lens lookups.
It can *never* be guaranteed that the return list of lenses of lensfun contains
only one entry, even if the lens model name was exactly correct.  Lenses with a
longer lens name (containing the shorter one) will also be returned.  However,
it is guaranteed that the exact match will be returned as the zeroth list
element.