Bug 9807: Correct Olympus Focus Distance
Originally the Focus Distance seems to be stored in a Rational (Fractional) format.

There seems to be some confusion about where the stored value is in centimeters versus millimeters.

This confusion seems to stem from the fact that some implementations seem to ignore the denominator of 10,
using the numerator directly (and incorrectly).

Exiv2's toFloat function properly divides the numerator by the denominator when outputting a Float (as centimeters).

So we need to divide by 100 to get meters.
(cherry picked from commit e32e81be92e8cfba9b728b272828da2173e8b3fc)