quick fix for #9949
this prevents a segmentation violation in the sharpen module for small images.
for #9949 this is only a fix of a symptom. the underlying root cause of that bug is
in the clipping module which does not take care that the clipping area needs to
lie within the image frame and should not have a width/height of zero