Add libsecret support
This patch adds the option to use libsecret as a password storage backend in
darktable. Because the gnome-keyring API has been deprecated and has been
replaced with the implementation of a secret service that can be accessed via
libsecret it is recommended to use libsecret. Since kwallet already supports the
libsecret API this patch disables the kwallet and gnome-keyring backend if
libsecret is available automatically. If libsecret is not available on the
system it automatically falls back to the gnome-keyring API and kwallet, if
those are available and enabled.

Since attributes are not stored in an encrypted way and therefore are not part
of the secret, this patch converts the passed attributes into the JSON format
and stores the data as the secret. Attributes are only used to lookup the
corresponding secret item correctly. To read the encrypted attributes this patch
then converts the stored JSON data into a hashtable with said attributes.