Markesteijn: clarify/clean-up min/max green calc
- Use 0.0f instead of 0 (which was an artifact from ushort-based dcraw
  code)
- use consts when possible
- add comments to clarify implementation

A hope of the comments is to give a rationale for using == to compare
two floats. The code works as written, but if it continues to look
worrisome, a workaround could be to create a separate "new_pair"
boolean rather than assigning 0.0f to max to signify a new pair.