Properly compute the buffer_size using the dt_mipmap_buffer_dsc size.
There was still 2 instances of hard coded 4*(uint32_t) for the size
of the header (dt_mipmap_buffer_dsc). But since change to use size_t
the buffer was bigger.

Fix a buffer overflow. With a bit of luck it should be the last
issue with the new mipmap implementation.