gphoto: only scan for devices when explicitly requested
over the years this has caused various issues in various situations,
sometimes even allowing darktable to not start because of lack of
a USB controller for example.

gphoto should never be put in a position where it can cause trouble
for users who are not using any of it's features.