Define DT_MAX_PATH_FOR_PARAMS
WARNING: it should *NEVER* be changed!
Else old presets (created with different PATH_MAX) will break.

We can not use e.g. PATH_MAX because because it is not guaranteed
to be same across all compilers and systems.

We could use _XOPEN_PATH_MAX which is defined as 1024 bytes
in the standard - any Unix(r) or Unix-alike which claims
adherence to  IEEE Std 1003.1(2004)

But i'm not sure whether we will/will not run on systems
that do/do not claim such adherence.