Prevent SIGFPE in slideshow if collection is empty
This patch resolves a floating point exception that occurs in

        int32_t rand = ran % cnt;

if the collection is empty and the variable cnt is therefore 0 which
results in a division by zero.

In addition a try_enter function is implemented to check the collection
count before trying to enter the slideshow view. If the collection is empty
a messages is displayed to tell the user so.