Make shooting in tether mode work
When using the shutter button on the camera dt should react by
downloading the image and importing it. However, there was no listener
by default which did that, so the images ended up under /tmp/ and didn't
get imported. No idea since when that was the case. I hope that it works
now.
Caveat: This has the slight chance of a race condition when having the
import from camera-dialog open while in tethering mode, since the
previews grabbed from the camera could end up bein treated as images. At
least that's my theory. In practice it should not be possible to have
the dialog open while in tethering mode so I don't care too much for
now.