Redo the image layout.
The image layout is now done by dt and not by CUPS. This give us better
control and work around the limitation where the margins are not taken
into account by some printers (as documented, but it is still strange
that some printers are behaving correctly).

This make the CUPS code cleaner and lighter as many options setting
have been removed.

A new format (not user visible) has been introduced and is used exclusively
by the print module. This format is TIFF 8 or 16 bits and can append white
margins to the image to ensure proper layout on the page.

The TIFF format has been reverted to its original implementation and is
not used anymore by the print module.

This print format could be enhanced later to do some sharpening or apply
directly the printer profile.