cache: notify everybody each time a write lock is dropped.
should fix some lighttable exposure problems. this is now necessary
because the new cache returns 0 in testget() even when the image is
there, but is currently write locked (was previously blocking then).
this behaviour in turn is necessary with the non-lock-free hashtable
we're now using, we would block a lot more than before now.