Change order of the profil/intent combo in export dialog.
The intent depends on the profil, it seems better to have it before.
In books or paper the profile is always described before and then the
intent apply for this profile is discussed. So again, the new
order seems more logical (and match what is done in print module).

Adapted from af2b5c0c9141a0cc806bb80950365429e954ad12 by Pascal Obry