new module proposal "color reconstruction"
this modules tries to guess the right color of overexposed highlights based
on the colors of nearby well exposed areas. the underlying algorithm uses a
bilateral filter. this module acts in Lab color space. it works alone but can
also take benefit of the "reconstruct color" option in the highlight recovery module.