opencl: de-activate opencl in case GPU is slower than CPU
when darktable sees an new config file or if the OpenCL setup (GPUs) has changed
we do a simple benchmarking test. If the fastest GPU is not clearly ahead of the
CPU we de-activate OpenCL. user can always activate it again in the preferences
dialog. rationale: avoid that a slow GPU negatively affects experience of users
who are unaware of that fact.