Add missing variables to OpenMP pragmas
Intel's ICC 15.0.2 errors out when the parallelized code accesses
variables whose sharing approach is not properly defined (especially
with a default(none)).

It's a question which sharing mode to use here, though. GCC apparently
defaults to shared, but it complains loudly when the code defines this
explicitly ("error: ‘max’ is predetermined ‘shared’ for ‘shared’").
Random Googling leads to a performance optimization guide from Oracle
[1] which apparently suggests to use "firstprivate" for const data,
although with questionable reasoning ("If a SHARED variable in a
parallel region is read by the threads executing the region, but not
written to by any of the threads, then specify that variable to be
FIRSTPRIVATE instead of SHARED. This avoids accessing the variable by
dereferencing a pointer, and avoids cache conflicts."  -- what cache
conflicts does it talk about?).

[1] http://docs.oracle.com/cd/E19059-01/stud.10/819-0501/7_tuning.html